+++ packages/koop-output-geoservices/index.js
@@ -10,7 +10,36 @@ Geoservices.prototype.featureServer = function (req, res) {
   })
 }
 
+Geoservices.prototype.featureServerRestInfo = function (req, res) {
+  FeatureServer.route(req, res)
+}
+
 Geoservices.routes = [
+  {
+    path: '$namespace/rest/info',
+    methods: ['get', 'post'],
+    handler: 'featureServerRestInfo'
+  },
+  {
+    path: '$namespace/rest/services/$providerParams/FeatureServer/:layer/:method',
+    methods: ['get', 'post'],
+    handler: 'featureServer'
+  },
+  {
+    path: '$namespace/rest/services/$providerParams/FeatureServer/layers',
+    methods: ['get', 'post'],
+    handler: 'featureServer'
+  },
+  {
+    path: '$namespace/rest/services/$providerParams/FeatureServer/:layer',
+    methods: ['get', 'post'],
+    handler: 'featureServer'
+  },
+  {
+    path: '$namespace/rest/services/$providerParams/FeatureServer',
+    methods: ['get', 'post'],
+    handler: 'featureServer'
+  },
   {
     path: 'FeatureServer/:layer/:method',
     methods: ['get', 'post'],